Edited the Readme and replaced the old main.py by the new implementation for packaging porpuses
