+++ openweathermap/main.py
@@ -17,7 +17,7 @@ class OpenWeatherMap(OMPluginBase):
     """
 
     name = 'OpenWeatherMap'
-    version = '1.0.2'
+    version = '2.0.0'
     interfaces = [('config', '1.0')]
 
     config_description = [{'name': 'api_key',
@@ -29,16 +29,14 @@ class OpenWeatherMap(OMPluginBase):
                           {'name': 'lng',
                            'type': 'str',
                            'description': 'A location longitude which will be passed to OpenWeatherMap.'},
-                          {'name': 'main_mapping',
-                           'type': 'section',
-                           'description': 'Mapping betweet OpenMotics Virtual Sensors and OpenWeatherMap forecasts. See README.',
-                           'repeat': True,
-                           'min': 0,
-                           'content': [{'name': 'sensor_id', 'type': 'int'},
-                                       {'name': 'time_offset', 'type': 'int'}]},
+                          {'name': 'time_offset',
+                           'type': 'int',
+                           'description': 'Time offset for forecast in minutes (only positive int values)'
+                           },
+                            # Not implemented yet, set to -1 in the meanwhile
                           {'name': 'uv_sensor_id',
                            'type': 'int',
-                           'description': 'Sensor ID for storing the UV index (the UV index will be set as temperature). -1 if not needed.'}]
+                           'description': 'Not yet implemented, set to -1\nSensor ID for storing the UV index (the UV index will be set as temperature). -1 if not needed.'}]
 
     default_config = {'api_key': ''}
 
@@ -55,14 +53,15 @@ class OpenWeatherMap(OMPluginBase):
 
         self._previous_output_state = {}
 
+        self.connector.sensor.subscribe_status_event(OpenWeatherMap.handle_sensor_status, version=2)
+        self._sensor_dto = None
+
         logger.info("Started OpenWeatherMap plugin")
 
     def _read_config(self):
         self._api_key = self._config.get('api_key', '')
 
-        main_mapping = self._config.get('main_mapping', [])
-        self._current_mapping = [entry for entry in main_mapping if entry['time_offset'] == 0]
-        self._forecast_mapping = [entry for entry in main_mapping if entry['time_offset'] > 0]
+        self._time_offset = self._config.get('time_offset')
         self._uv_sensor_id = int(self._config.get('uv_sensor_id', -1))
 
         self._uv_endpoint = 'http://api.openweathermap.org/v3/uvi/{lat},{lon}/{date}Z.json?appid={api_key}'
@@ -85,36 +84,36 @@ class OpenWeatherMap(OMPluginBase):
     def run(self):
         previous_values = {}
         accuracy = 5
+        if self._sensor_dto == None:
+            self._register_sensor()
         while True:
             if self._enabled:
                 start = time.time()
                 sensor_values = {}
                 calls = 0
-                if len(self._forecast_mapping) > 0:
+                if self._time_offset > 0:
                     try:
                         calls += 1
                         response = requests.get(url=self._forecast_endpoint.format(lat=self._latitude,
-                                                                                   lon=self._longitude,
-                                                                                   api_key=self._api_key),
+                                                                                    lon=self._longitude,
+                                                                                    api_key=self._api_key),
                                                 headers=self._headers)
                         if response.status_code != 200:
                             logger.error('Forecast call failed: {0}'.format(response.json()['message']))
                         else:
                             result = response.json()['list']
-                            for sensor in self._forecast_mapping:
-                                sensor_id = sensor['sensor_id']
-                                wanted_time = start + (sensor['time_offset'] * 60)
-                                selected_entry = None
-                                for entry in result:
-                                    if selected_entry is None or abs(entry['dt'] - wanted_time) < abs(selected_entry['dt'] - wanted_time):
-                                        selected_entry = entry
-                                if selected_entry is None:
-                                    logger.error('Could not find forecast for virtual sensor {0}'.format(sensor_id))
-                                    continue
-                                sensor_values[sensor_id] = [selected_entry['main']['temp'], selected_entry['main']['humidity'], None]
+                            wanted_time = start + (self._time_offset * 60)
+                            selected_entry = None
+                            for entry in result:
+                                if selected_entry is None or abs(entry['dt'] - wanted_time) < abs(selected_entry['dt'] - wanted_time):
+                                    selected_entry = entry
+                            if selected_entry is None:
+                                logger.error('Could not find forecast for virtual sensor {0}'.format(sensor_id))
+                                continue
+                            sensor_values[0] = [selected_entry['main']['temp'], selected_entry['main']['humidity'], None]
                     except Exception as ex:
                         logger.exception('Error while fetching forecast temperatures')
-                if len(self._current_mapping) > 0:
+                elif self._time_offset == 0:
                     try:
                         calls += 1
                         response = requests.get(url=self._current_endpoint.format(lat=self._latitude,
@@ -125,11 +124,10 @@ class OpenWeatherMap(OMPluginBase):
                             logger.error('Current weather call failed: {0}'.format(response.json()['message']))
                         else:
                             result = response.json()
-                            for sensor in self._current_mapping:
-                                sensor_id = sensor['sensor_id']
-                                sensor_values[sensor_id] = [result['main']['temp'], result['main']['humidity'], None]
+                            sensor_values[0] = [result['main']['temp'], result['main']['humidity'], None]
                     except Exception as ex:
                         logger.exception('Error while fetching current temperatures')
+                # Currently there is no register_brightness_... in the sensor connector
                 if 0 <= self._uv_sensor_id <= 31:
                     try:
                         execute = True
@@ -158,19 +156,20 @@ class OpenWeatherMap(OMPluginBase):
                                 execute = False
                     except Exception as ex:
                         logger.exception('Error while fetching UV index')
-                # Push all sensor data
-                try:
-                    for sensor_id, values in sensor_values.items():
-                        if values != previous_values.get(sensor_id, []):
-                            logger.info('Updating sensor {0} to temp: {1}, humidity: {2}'.format(sensor_id,
-                                                                                                 values[0] if values[0] is not None else '-',
-                                                                                                 values[1] if values[1] is not None else '-'))
-                        previous_values[sensor_id] = values
-                        result = json.loads(self.webinterface.set_virtual_sensor(None, sensor_id, *values))
-                        if result['success'] is False:
-                            logger.error('Error when updating virtual sensor {0}: {1}'.format(sensor_id, result['msg']))
-                except Exception as ex:
-                    logger.exception('Error while setting virtual sensors')
+
+                # Push all sensor data                
+                for _, values in sensor_values.items():
+                    if values[0] != previous_values.get(0):
+                        logger.info('Updating sensor {0} to temp: {1}'.format(self._sensor_dto["name"],
+                                                                                                values[0] if values[0] is not None else '-'))
+                    previous_values[0] = values[0]  # Only temperature
+                    try:
+                        if self._sensor_dto:
+                            self.connector.sensor.report_state(sensor=self._sensor_dto,
+                                                            value=values[0])
+                    except Exception:
+                        logger.exception('Error while reporting sensor state')
+
                 # Wait a given amount of seconds
                 sleep = 60 * calls - (time.time() - start) + 1
                 if sleep < 0:
@@ -198,3 +197,31 @@ class OpenWeatherMap(OMPluginBase):
         self._read_config()
         self.write_config(config)
         return json.dumps({'success': True})
+
+    @staticmethod
+    def handle_sensor_status(event):
+        logger.info('Received sensor status from gateway: {0} {1}'.format(
+            event.data['id'], event.data['value']
+        ))
+
+    def _register_sensor(self):
+        logger.info('Registering Temperature sensor...')
+        try:
+            sensor = self.connector.sensor.register_temperature_celcius(external_id='222222',
+                                                                        name='OWM-temp-sensor')
+            logger.info('Registered %s' % sensor)
+            self._sensor_dto = sensor
+        except Exception:
+            logger.exception('Error registering sensor')
+            self._sensor_dto = None
+
+    # TODO: also register a humidity sensor and set it's value
+        # logger.info('Registering Humidity sensor...')
+        # try:
+        #     sensor = self.connector.sensor.register_humidity_percent(external_id='333333',
+        #                                                                 name='OWM-humidity-sensor')
+        #     logger.info('Registered %s' % sensor)
+        #     self._sensor_dto = sensor
+        # except Exception:
+        #     logger.exception('Error registering sensor')
+        #     self._sensor_dto = None
